refactor: Rename `from` and `to` to `fromDomain` and `toDomain` and move SplashScreenRepository
This commit renames the `from` and `to` functions in various UI model classes to `fromDomain` and `toDomain` respectively. This change improves clarity and consistency in mapping between domain and UI models.

Additionally, `SplashScreenRepository` has been moved from the `domain` package to the `data` package to better reflect its role in data handling.

Specific changes:
- Renamed `from()` to `fromDomain()` in `AnnouncementUi.kt`, `ForumUi.kt`, `RepliesUi.kt`, `RoadMapUi.kt`, and `UserUi.kt`.
- Renamed `to()` to `toDomain()` in `AnnouncementUi.kt`, `ForumUi.kt`, and `UserUi.kt`.
- Moved `SplashScreenRepository.kt` from `app/src/main/java/com/mycheva/app/splashscreen/domain` to `app/src/main/java/com/mycheva/app/splashscreen/data`.
- Updated import statements in `SplashScreenViewModel.kt` to reflect the new location of `SplashScreenRepository`.
- Updated calls to the renamed mapping functions in `AnnouncementViewModel.kt`, `ForumScreen.kt`, `RepliesScreen.kt`, and `RoadMapViewModel.kt`.
