+++ app/src/main/java/com/mycheva/app/forum/main/presentation/ForumScreen.kt
@@ -180,7 +180,7 @@ fun SharedTransitionScope.ForumScreen(
                             HorizontalDivider(modifier = Modifier.fillMaxWidth())
                         }
                         ForumPostCard(
-                            forum = ForumUi.fromForum(post),
+                            forum = ForumUi.fromDomain(post),
                             modifier = Modifier.padding(bottom = 16.dp),
                             onNavigate = {
                                 onNavigate(CommentScreen(forumId = it))
