+++ app/src/main/java/com/mycheva/app/splashscreen/presentation/SplashScreenViewModel.kt
@@ -2,7 +2,7 @@ package com.mycheva.app.splashscreen.presentation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.mycheva.app.splashscreen.domain.SplashScreenRepository
+import com.mycheva.app.splashscreen.data.SplashScreenRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
